% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_wt.R
\name{r_wt}
\alias{r_wt}
\title{Weighted Pearson's correlation}
\usage{
r_wt(y_h, y, wt, na.rm = FALSE)
}
\arguments{
\item{y_h}{Vector of predicted values.}

\item{y}{Vector of true values.}

\item{wt}{Vector of sample weights.}

\item{na.rm}{Logical whether to remove NA values.}
}
\description{
Calculate a weighted version of Pearson's correlation coefficient.
}
\details{
This approach relies on normalizing the weights to obtain "reliability weights", which sum to 1.
These are used to calculate the weighted covariance between predictions \code{y_h} and true values \code{y} using their weighted means (also calculated using the normalized weights).
The correlation is then calculated using the weighted covariance and the weighted variances for \code{y_h} and \code{y}.
}
\examples{
#generate random data
a <- runif(100, 0, 3)

#simulate a linear dependent covariate with noise
b <- a + rnorm(100)

#check unweighted correlation
cor(a, b)

#generate random vector of weights
wt = runif(100)

#return weighted correlation
r_wt(a, b, wt)

#compare to the cov.wt function
cov.wt(x = data.frame(a, b), wt = wt, cor = TRUE)$cor[2]

}
