% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_wt.R
\name{rmse_wt}
\alias{rmse_wt}
\title{Weighted Root Mean Squared Error}
\usage{
rmse_wt(y_h, y, wt)
}
\arguments{
\item{y_h}{Vector of predicted values.}

\item{y}{Vector of true values.}

\item{wt}{Vector of sample weights.}
}
\description{
Calculate weighted or unweighted root mean squared error.
}
\details{
Weighted RMSE is calculated by taking the root of the weighted MSE, as calualated using \code{mse_wt}.
}
\examples{
#generate random data
a <- runif(100, 0, 3)

#simulate a linear dependent covariate with noise
b <- a + rnorm(100)

#check unweighted RMSE
rmse(a, b)

#generate random vector of weights
wt = runif(100)

#return weighted MSE
rmse_wt(a, b, wt)

}
