% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFunc.R
\name{varFunc}
\alias{varFunc}
\title{Variogram model functions}
\usage{
varFunc(h, a, fun)
}
\arguments{
\item{h}{Vector of distances.}

\item{a}{Numerical. Fitted range parameter from a variogram model.}
}
\description{
Functions to transform distance into semivariance.
}
\details{
Transforms a vector of distances \code{h} to semivariance according to a semivariogram model function. Note this does not include a nugget component, which should be added to the output of these functions if desired.
}
\examples{

}
