% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ve_wt.R
\name{ve}
\alias{ve}
\title{Weighted Variance Explained}
\usage{
ve(y_h, y)
}
\arguments{
\item{y_h}{Vector of predicted values.}

\item{y}{Vector of true values.}

\item{wt}{Vector of sample weights.}
}
\description{
Calculate weighted or unweighted variance explained.
}
\details{
Unweighted variance explained is calculated by dividing the residual sum of squares between \code{y_h} and \code{y} from the total sum of squares of \code{y}, and subtracting from 1.
Weighted variance is obtained by weighting the residual and total sum of squares by the vector \code{wt} prior to dividing.
}
\examples{
#generate random data
a <- runif(100, 0, 3)

#simulate a linear dependent covariate with noise
b <- a + rnorm(100)

#check unweighted VE
ve(a, b)

#generate random vector of weights
wt = runif(100)

#return weighted VE
ve_wt(a, b, wt)

}
