% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagger.R
\name{bagger}
\alias{bagger}
\title{Bagger}
\usage{
bagger(
  y,
  x,
  data,
  p = rep(1, nrow(data)),
  mtry = length(x),
  mtry_replace = FALSE
)
}
\arguments{
\item{y}{Character naming the response variable.}

\item{x}{Character vector naming the predictor variables.}

\item{data}{Data frame containing observations of \code{x} and \code{y}.}

\item{p}{Optional vector of probabilities for drawing the observations.}

\item{mtry}{Integer indicating how many predictors to sample from \code{x}. Default is all.}

\item{mtry_replace}{Logical whether to sample the predictors with replacement, is \code{mtry} is less than the the number of predictors indicated in \code{x}.}
}
\value{
A list containing the bootstrapped sample data \verb{$data}, a vector of indices indicating which samples were drawn during bootstrapping \verb{$oob},
and a vector of which predictor variables were selected \verb{$x}.
}
\description{
A generic function for drawing bagging samples for machine learning models.
}
\details{
Draws bootstrap samples (with replacement) from a data frame, allowing for optional specification of probability weights and feature bagging (random selection of predictors).
}
\examples{
data(iris)
#define the vector of predictors to include
x <- c('Sepal.Length', 'Sepal.Width', 'Petal.Length', 'Petal.Width')

b <- bagger(y = 'Species', x, data = iris)

#compare the number of cases to the original data
table(iris$Species)
table(b$data$Species)

#down-weight the probability of drawing "setosa" observations
iris$prob <- 1
iris$prob[iris$Species == 'setosa'] <- 0.5

#re-draw using the probabilities
b <- bagger(y = 'Species', x, data = iris, p = iris$prob)

table(iris$Species)
table(b$data$Species)

}
