% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagger.R
\name{bagger}
\alias{bagger}
\title{Bagger}
\usage{
bagger(
  y,
  x,
  data,
  p = rep(1, nrow(data)),
  mtry = length(x),
  mtry_replace = FALSE,
  model,
  model_params = NULL,
  n = 500
)
}
\arguments{
\item{y}{Character naming the response variable.}

\item{x}{Character vector naming the predictor variables.}

\item{data}{Data frame containing observations of \code{x} and \code{y}.}

\item{p}{Optional vector of probabilities for drawing the observations.}

\item{mtry}{Integer indicating how many predictors to sample from \code{x}. Default is all.}

\item{mtry_replace}{Logical whether to sample the predictors with replacement, is \code{mtry} is less than the the number of predictors indicated in \code{x}.}

\item{model}{}
}
\value{
List containing the models in the ensemble \verb{$models}, and the out-of-bag predictions \verb{$oob_preds}.
}
\description{
A generic function for creating a bagging model ensemble.
}
\details{

}
\examples{

}
