% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covWt.R
\name{covWt}
\alias{covWt}
\title{covWt}
\usage{
covWt(x, model, m, a, nug, psil, dmat)
}
\arguments{
\item{model}{variogramModel class, for example as returned from \code{\link[gstat:fit.variogram]{gstat::fit.variogram()}} or \code{\link[automap:autofitVariogram]{automap::autofitVariogram()}}.}

\item{m}{Character. Type of variogram model. Must be one of "Exp", "Gau", or "Sph". Ignored if \code{model} is provided.}

\item{a}{Numeric. Range parameter of a variogram model. Ignored if \code{model} is provided.}

\item{nug}{Numeric. Nugget parameter of a variogram model. Ignored if \code{model} is provided.}

\item{psil}{Numeric. Partial sill of a variogram model. Ignored if \code{model} is provided.}

\item{dmat}{Distance matrix. Can be of class "dist" or a symmetric matrix of distances between points with diagonals included.}
}
\value{
A vector of weights.
}
\description{
Calculate covariance weighting for spatial data points based on residual spatial autocorrelation.
}
\details{
Calculates spatial data point weights based on the sum of residual spatial covariances for each point.
Provide either a variogram model from \link{gstat} or the parameters of a model fitted to prediction residuals.
Currently supports exponential ("Exp"), Gaussian ("Gau"), and spherical ("Sph") models.
Currently does not support nested variogram models. If you would like to use these, call \code{covWt} separately for each model and sum them.
}
\examples{


}
