% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFunc.R
\name{varFunc}
\alias{varFunc}
\title{Variogram model functions}
\usage{
varFunc(h, a, fun)
}
\arguments{
\item{h}{Vector of distances.}

\item{a}{Numerical. Fitted range parameter from a variogram model.}

\item{fun}{Character string indicating the variogram model function. Must be one of "Exp", "Gau", or "Sph".}
}
\value{
A vector of semivariance values.
}
\description{
Functions to transform distance into semivariance.
}
\details{
Transforms a vector of distances \code{h} to semivariance $\gamma(h)$ according to a semivariogram model function. Note this does not include nugget or sill components
Use $c(0)+c(varFunc())$, for example, to plot semivariance, where $c(0)$ is the nugget and $c$ is the partial sill.
Currently supported model functions are the exponential, Gaussian, and spherical.
}
\examples{
#set a range parameter (e.g., taken from gstat::fit.variogram())
r = 250

#generate some distances to convert to semivariances
d <- seq(0, 1000, 50)

varFunc(h=d, a=r, fun = 'Exp')

#look at plots of the different functions
plot(varFunc(h=d, a=r, fun = 'Exp'))
points(varFunc(h=d, a=r, fun = 'Gau'), col = 'red')
points(varFunc(h=d, a=r, fun = 'Sph'), col = 'blue')

}
